requirements.txt
google-generativeai
pdf2image
pytesseract
pandas
reportlab
markdown
python-docx
docxtpl